refactor: update navbar using react-bootstrap
