--- frontend/src/components/Navbar.jsx
@@ -1,54 +1,0 @@
-import { Link, useNavigate } from 'react-router-dom';
-import { useAuth } from '../context/AuthContext';
-
-const Navbar = () => {
-  const { user, logout } = useAuth();
-  const navigate = useNavigate();
-
-  const handleLogout = () => {
-    logout();
-    navigate('/login');
-  };
-
-  return (
-    <nav className="bg-blue-600 text-white p-4 flex justify-between items-center">
-      <Link to="/" className="text-2xl font-bold">Startup Incubation</Link>
-      <div>
-        {user ? (
-          <>
-            <Link to="/applications" className="mr-4">Applications</Link>
-            <Link to="/startup" className="mr-4">Startup</Link>
-            <Link to="/profile" className="mr-4">Profile</Link>
-            
-            <Link to="/applications/apply" 
-              className="bg-green-500 px-4 py-2 mx-2 rounded hover:bg-green-700">
-                Apply for Program
-            </Link> 
-            <Link
-              onClick={handleLogout}
-              className="bg-red-500 px-4 py-2 rounded hover:bg-red-700"
-            >
-              Logout
-            </Link>
-          </>
-        ) : (
-          <>
-            <Link to="/applications/apply" 
-              className="bg-green-500 px-4 py-2 mx-2 rounded hover:bg-green-700">
-                Apply for Program
-            </Link> 
-            <Link to="/login" className="mr-4">Login</Link>
-            <Link
-              to="/register"
-              className="bg-gray-500 px-4 py-2 rounded hover:bg-gray-700"
-            >
-              Register
-            </Link>
-          </>
-        )}
-      </div>
-    </nav>
-  );
-};
-
-export default Navbar;
